requirements.txt
agdistispy==1.0.5
annoy==1.15.1
boto==2.49.0
boto3==1.9.115
botocore==1.12.115
bpemb==0.2.11
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Deprecated==1.2.5
docutils==0.14
fasteners==0.14.1
flair==0.4.1
flashtext==2.7
future==0.17.1
gensim==3.7.1
h5py==2.8.0
hyperopt==0.1.2
idna==2.8
jieba==0.39
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.0.1
lmdb==0.94
lz4==2.1.2
marisa-trie==0.7.5
matplotlib==3.0.3
monotonic==1.5
mpld3==0.3
mwparserfromhell==0.5.2
networkx==2.2
numpy==1.16.2
pymagnitude==0.1.120
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.1
regex==2018.1.10
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
segtok==1.5.7
sentencepiece==0.1.8
simstring-pure==1.0.0
six==1.12.0
sklearn==0.0
smart-open==1.8.0
sqlitedict==1.6.0
tabulate==0.8.3
torch==1.0.1.post2
tqdm==4.31.1
Unidecode==1.0.23
urllib3==1.24.1
wikipedia2vec==1.0.3
wrapt==1.11.1
xxhash==1.3.0
